requirements.txt
accelerate==1.12.0
addict==2.4.0
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aiohttp-retry==2.8.3
aioice==0.10.2
aiortc==1.14.0
aiosignal==1.4.0
aiosqlite==0.22.1
annotated-types==0.7.0
anthropic==0.49.0
anyascii==0.3.3
anyio==4.12.1
APScheduler==3.11.2
asyncua==1.1.8
attrs==25.4.0
av==16.1.0
backoff==2.2.1
boto3==1.41.5
botocore==1.41.5
cachetools==5.5.2
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
coloredlogs==15.0.1
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
Cython==3.0.12
dataclasses-json==0.6.7
defusedxml==0.7.1
distro==1.9.0
dnspython==2.8.0
docker==7.1.0
docopt==0.6.2
easyocr==1.7.2
einops==0.8.1
fastapi==0.115.14
filelock==3.17.0
filetype==1.2.0
flatbuffers==25.12.19
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2026.1.0
ftfy==6.3.1
google-crc32c==1.8.0
h11==0.16.0
h5py==3.15.1
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.7
ifaddr==0.2.0
ImageIO==2.37.2
inference==0.63.5
inference-exp==0.17.3
inference-sdk==0.63.5
iniconfig==2.3.0
isodate==0.7.2
Jinja2==3.1.6
jiter==0.12.0
jmespath==1.0.1
kiwisolver==1.4.9
langdetect==1.0.9
lazy_loader==0.4
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib==3.10.8
mdurl==0.1.2
ml_dtypes==0.5.4
MouseInfo==0.1.3
mpmath==1.3.0
mss==10.1.0
multidict==6.7.0
mypy_extensions==1.1.0
networkx==3.6.1
ninja==1.13.0
num2words==0.5.14
numpy==2.3.5
nvidia-ml-py==12.575.51
onnx==1.20.1
onnxruntime==1.21.1
onvif-zeep-async==2.0.0
openai==1.109.1
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.2
paho-mqtt==1.6.1
pandas==2.3.3
peft==0.17.1
pi_heif==1.1.1
pillow==11.3.0
pillow-avif-plugin==1.5.2
platformdirs==4.5.1
pluggy==1.6.0
prometheus-fastapi-instrumentator==6.0.0
prometheus_client==0.24.1
propcache==0.4.1
protobuf==6.33.4
psutil==7.2.1
py-cpuinfo==9.0.0
PyAutoGUI==0.9.54
pybase64==1.0.2
pyclipper==1.4.0
pycocotools==2.0.11
pycparser==2.23
pydantic==2.11.10
pydantic-settings==2.7.1
pydantic_core==2.33.2
pydot==2.0.0
pyee==13.0.0
PyGetWindow==0.0.9
Pygments==2.19.2
PyJWT==2.10.1
pylibsrtp==1.0.0
pylogix==1.0.5
pymodbus==3.8.3
PyMsgBox==2.0.1
pyobjc-core==12.1
pyobjc-framework-Cocoa==12.1
pyobjc-framework-Quartz==12.1
pyOpenSSL==25.3.0
pyparsing==3.3.1
pypdfium2==4.30.0
pyperclip==1.11.0
PyRect==0.2.0
PyScreeze==1.0.1
pytest==8.4.2
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-doctr==0.11.0
python-dotenv==1.0.1
pytweening==1.2.0
pytz==2025.2
pyvips==2.2.3
PyYAML==6.0.3
qrcode==8.0
RapidFuzz==3.14.3
redis==5.0.8
regex==2026.1.15
requests==2.32.5
requests-file==3.0.1
requests-toolbelt==1.0.0
rf-clip==1.1
rf_groundingdino==0.3.0
rich==14.2.0
roboflow==1.2.11
rubicon-objc==0.5.3
s3transfer==0.15.0
safetensors==0.7.0
scikit-image==0.25.2
scipy==1.17.0
segmentation_models_pytorch==0.5.0
sentencepiece==0.2.1
setuptools==80.9.0
shapely==2.0.7
shellingham==1.5.4
simple-pid==2.0.1
six==1.17.0
slack_sdk==3.33.5
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.46.2
structlog==24.4.0
supervision==0.27.0
sympy==1.14.0
tifffile==2026.1.14
timm==1.0.24
tldextract==5.1.3
tokenizers==0.22.2
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
transformers==4.57.5
twilio==9.3.8
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.12.2
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
wcwidth==0.2.14
wheel==0.45.1
yapf==0.43.0
yarl==1.22.0
zeep==4.3.2
zxing-cpp==2.2.0
